Bug#32106582 no ciphers available when connecting
Problem
=======

When opening a TLS connection over classic protocol a client sends
a short client greeting which only contains the TLS-capability.

Router was sending a full client greeting with the TLS capability
set which sometimes confused the server side which treated the extra
bytes as a TLS greeting packet.

Change
======

- send a short Client Greeting packet.

RB: 25438