Bug#31997733 SELECT COUNT(*) WORKS VERY SLOW FOR UNCOMMITTED DATA
Issue
=====
In method Parallel_reader::Ctx::traverse_recs, for every record traversed we
get the visible version of the record for the transaction and if the record is
visible we check if we've reached the end of range of the scan context. If many
records are not visible we might pass the endpoint and the scan will continue
until the next visible record. What this means is in the extreme case where
none of the records in a table are visible the scan will not stop until end of
table index wasting IO.

Fix
===
While checking if we've reached the end range check the latest version of the
record instead of the version visible to the transaction.

This patch is based on a contribution from Brian Yue.

RB: 25409
Reviewed-by: Sunny Bains <Sunny.Bains@oracle.com>