BUG#32068730: FAILURES ON GR_AUTOREJOIN_WHILE_RECOVERY_LOCKED AND GR_AUTOREJOIN_WITH_CLONE_WHI
Tests
  group_replication.gr_autorejoin_while_recovery_locked and
  group_replication.gr_autorejoin_with_clone_while_recovery_locked
are failing on el6-x86-32bit platform on PB2.
More precisely, the tests are timing out.

After analysis it was discovered that the root cause was the GR
rejoin thread not being able to stop the recovery channel.
This was happening because the thd of the rejoin thread was being
instantiated on the object stack instead of the local stack.

To fix the above issue, now the rejoin thread instantiates its thd
on the local stack.

RB: 25424