Bug#32130025: OPT_TRACE-T COULD USE MYSQL'S BUILT-IN JSON PARSER FOR VERIFICATION
The opt_trace-t unit test is changed in the following ways:

Use MySQL's built-in JSON parser for verifying that the generated
trace is a valid JSON document. Remove the old verification code,
which sent the trace to a Python program.

Enable verification of the generated trace by default, since the
verification no longer relies on external programs.

Remove testing of the add_hex() function, since it generates invalid
JSON. Both the old Python-based verifier and the new verifier say that
it is invalid. The function is only used by the unit test, so it is
also removed from the Opt_trace interface.

Change a test case which creates a nested JSON object structure with
depth 101, so that it instead creates a structure with depth 100. The
reason is that the built-in JSON parser rejects documents that have
depth greater than 100. The test case still exercises the code path
that it is supposed to test.

Remove the SetUpTestCase() function. It changes the system character
set to UTF-8, which has been the default for a long time, so it is no
longer needed.

Change-Id: I56b93494b294836961218cfe76f61255a723f517