MEB.PAGE_COMPRESSION TEST FAILURE WITH BUG#31744694 FIX
Issue
=====
In the case of import tables, at a later stage, we try to set the compression
type in the fil_space_t object to enable pages to be compressed at the time of
flush after the import.

As part of the BUG#31744694 fix the import code was made to abort in case of a
failure while setting this compression type. As a result of this, on systems
that do not support punch hole, the MEB test would fail because of this
early abort.

Fix
===
Compression during import should be a best effort scenario. We should
try setting the compression type and move forward regardless of whether
we were successful or not as we should try to successfully import tables
first.

The patch removes the abort code in case of a failure while setting the
compression type and throws a client warning instead.

RB: 25388
Reviewed-by: Mayank Prasad <mayank.prasad@oracle.com>