Bug#32149390 MAKE MTR SHUTDOWN THE SERVER THROUGH THE SOCKET/PIPE, NOT TCP
Problem: after the push of

Bug#31975014 SAFE_PROCESS ON WINDOWS DOES NOT WAIT FOR ALL CHILDREN TO TERMINATE
Bug#32084908 MTR FORCEFUL KILL ERROR IS NOT PROVIDING ANY INFO TO CORRELATE TEST

MTR now reports on processes that it is not able to shut down properly
after a test is finished.

Shutdown is done through mysqladmin over TCP.

There are reports on hanging processes from several tests in the test
suite.

However, in most of the cases, the servers are not hanging, they just
refuse to speak over TCP to mysqladmin.

Solution: change MTR to make mysqladmin speak on the socket (or pipe
in the case of Windows), just like mysqltest does.

Change-Id: Ic22e3b190fd96d7f52e9371edb8119e7257f4d78