Bug#32113020: OPTIMIZER TRACE STILL PRINTS FLOATING POINT NUMBERS IMPRECISELY [2/2, code]
Print floating point numbers with higher precision in optimizer trace.

Previously, they were printed with maximum six characters, which meant
that the precision would be very low for many values, since sign,
decimal point and exponent could take many of those characters. This
was especially bad for large numbers, which could have as little
precision as 1, and which could get rounded to values that exceeded
+/-DBL_MAX and would be rejected by JSON parsers.

Now they are printed with a precision of 6, also for large floating
point values.

Change-Id: Ib7cece1ddd341fe64ccf4e0d4fb93ec2faf622d5