BUG#32141729: ASSERTION FAILS WHEN SLAVE THREADS ARE TERMINATED AND ACF IS ENABLED
On Asynchronous Replication Connection Failover feature, when the
last channel with SOURCE_CONNECTION_AUTO_FAILOVER=1 is stopped, the
Monitor IO Thread is also stopped.
Though if the channel with SOURCE_CONNECTION_AUTO_FAILOVER=1 was on
error state the Monitor IO Thread was not stopped.

To solve the above issue, now failed IO Threads are also considered
on the conditions to stop the Monitor IO Thread.

RB: 25457