WL #14421: Recursive CTE in the hypergraph join optimizer
Implement recursive CTE in the hypergrah join optimizer. This is pretty
much the same as any other table, with the main exception that recursive
references are FOLLOW_TAIL paths instead of regular table scans,
and need to be on the very left of the join tree.

No unit test, as CTE necessarily needs to involve temporary tables,
which are hard to set up in a unit test.

Change-Id: I4abc8102559adfc0558bfc84a2b770a2d8ce1264