WL #14323: Hypergraph optimizer conflict detector [8/8, STRAIGHT_JOIN]
Respect STRAIGHT_JOIN instead of just treating it as any other inner
join.

The manual describes STRAIGHT_JOIN as: “STRAIGHT_JOIN is similar to
JOIN, except that the left table is always read before the right table.”
Thus, we mark it as noncommutative but otherwise identical to a normal
inner join. In particular, we still regard it as associative by the same
rules a normal inner join would be.

Change-Id: I16f4e065d6c01c93288cb3fbbbca78962fb2f488