WL#14189: Replace old terms in CHANGE MASTER TO and START SLAVE params
This patch implements aliases for CHANGE MASTER and its parameters
that have MASTER in their name. CHANGE MASTER has an alias CHANGE
REPLICATION SOURCE and the parameters that have MASTER in the name
have aliases that replace MASTER with SOURCE. There are also aliases
for START REPLICA parameters that follow the same rule. If the old
terminology is used, deprecation warnings are emitted.

Reviewed-by: Sven Sandberg <sven.sandberg@oracle.com>
Reviewed-by: Pedro Figueiredo <pedro.figueiredo@oracle.com>
Reviewed-by: Gleb Shchepa <gleb.shchepa@oracle.com>