BUG#32146931: usageCount on table has to be maintained while waiting for disk page retrievals
Reviewed-by: Mauritz Sundell <mauritz.sundell@oracle.com>

When calling disk_page_abort_prealloc we ignore the callback from this function
and avoid waiting for it and thus proceed with removal of the operation record
for the LQHKEYREQ operation. This leaves the table subject to removal before the
callback has completed which causes PGMAN to crash when the page is retrieved
from disk.

To avoid this we add an extra usageCount on the table especially for this
page cache miss that is decremented as soon as the page cache miss returns.
This means that we guarantee that the table is still there when we return
from the disk read.