Bug #30521812 - FAILURE IN DBDIH FOR TESTDICT -L 1 -N SCHEMATRANS
Background: When a table creation schema transaction (TCST) is
prepared, the table is in TS_CREATING state and it goes to TS_ACTIVE
state when the TCST commits at DIH block.

Scenario: The test crashes the DIH-coordinator node when the TCST is
committing.

Another node starts taking over for the coordinator. As a part of
handling the node failure :
a) DICT rolls the TCST forward and commits, resulting in the
   involved table to become TS_ACTIVE.
b) DIH starts removing the failed node from tables by moving
   ACTIVE-table replicas of the failed node from storedReplica
   to the oldStoredReplica list.

a) and b) are performed asynchronously in many rounds with
continueB. Interleaving a) and b) may cause a race. As a result, the
list (oldStoredRepls or storedRepls) where a failed node's replica is
residing becomes non-deterministic and differs between the recovering
node/new coordinator and the other DIH-participant nodes. This
difference violates the requirement of which list the failed node's
replicas can be found during the phases of the failed node recovery at
the other participants.

Fix: b) should cover not only tables in TS_ACTIVE, but also those in
TS_CREATING (prepared) state, since the prepared TCST will be rolled
forward anyway.

Also set the state of an aborting TCST from TS_CREATING to TS_DROPPING
to avoid any race there.

Reviewed by: Frazer Clement <frazer.clement@oracle.com>