Bug#32164436 linux-epoll-io-service drops events
Problem
=======

linux_epoll_io_service manages the wait-interest for file-descriptors
via the epoll_ctl()/epoll_wait() functions on Linux.

It uses oneshot events which remove themself from the epoll once an
event fired.

If a fd was waiting for more than one event (IN + OUT), epoll_ctl()
would remove the interest for IN and OUT event if only IN was
triggered and not report any further events while the
linux_epoll_io_service would still think the event is waited for.

Change
======

If multiple events are waited for and only one event triggers, add back
the rest of the events to epoll.

RB: 25490