Bug#25840784: PARTIAL UPDATE ON VIEWS DOES NOT WORK AS EXPECTED ON BASE TABLES
When updating a JSON value using JSON_SET, JSON_REPLACE or
JSON_REMOVE, the target column can sometimes be updated in-place. This
currently happens only when the target table of the update operation
is a base table. If the target table is an updatable view, the update
is always done by writing the full JSON value.

This patch enables in-place updates also when using updatable views,
by peeling off the Item_view_ref that wraps the Item_field pointing to
the base table.

Change-Id: I7c0ad3ea58cc9a458fac6248441fdca6788541c4