Bug#29800785 CLANG-FORMAT ABSTRACT_QUERY_PLAN
The files abstract_query_plan.cc and abstract_query_plan.h has
been left out of clang-formatting with justification that it's
part of NDB.

It might be true that ndbcluster plugin is the only user of
the interface exposed by abstract_query_plan, but the files
are compiled as part of MySQL Server and should thus adhere
to the same rules as rest of MySQL Server. This is difference
which just leads to confusion about which coding style should
be used for these files.

Fix by turning on on clang formatting of
sql/abstract_query_plan files.
Remove the exemption from .gitattributes

Change-Id: Id524794f6d69303f47f097a4a5af13ecd65e924f