Bug#32145233: DIFFS FROM MTR TESTS WITH --HYPERGRAPH
Post-push fix for a failure in main.outfile on some platforms with an
older version of the diff tool. The result file contains binary
characters which confuse the logic that detects if a query has been
skipped because it was not supported by the hypergraph join optimizer.
Now instead use an optimizer hint to disable the hypergraph optimizer
for the problematic query.

Change-Id: I2b97d16aed93267ddb890d587bc5ad58ba30c8b3