Bug#32139949 - MEB_DDL.CREATE_POS2 FAILS ON PB2 INTERMITTENTLY
Sporadically the server error log contains messages about pages being
"in the future".

This comes from redo log being restored incorrectly. When normal
incremental backups are restored, a wrong decision can be made and the
log from the incremental backup is appended to the log file for the
server restart, although the logs from the base backup and the
incremental backup may not be contiguous. This jump in log block numbers
lets the server stop to apply the log until its end, resulting in a
lower "current LSN" than some pages may have. Hence the pages appear to
be "in the future".

The wrong decision is made when the last checkpoint LSN is less than the
backup's start LSN. This can happen in relatively short tests when no
checkpoint is generated after the one, that the base backup has seen.
The incremental start LSN is 1 + the base backup's end LSN.

A normal incremental backup does always copy redo log from the log
block, that contains the latest checkpoint (even when this produces an
overlap in the logs). The server start needs the log segment from the
last checkpoint until the backup's end LSN. So it is safe to replace the
previously generated server log files.

However an incremental backup "with redo log only" copies the log only
from its start LSN, regardless whether the latest checkpoint is before
that. The copied log is always assured to be contiguous with the base
backup. In such case where the incremental backup does not contain a
checkpoint at all, we must not replace the previously generated server
log files, but append to them.

Additionally improve the logging of LSNs to allow quicker analysis in
the future.

To allow the the redo log writer to write the correct consistency LSN to
the log end mark, update it in the meta data before requesting log copy
completion.

Also for improved future analysis log the values found in the
backup_variables.txt files, that are used by an operation.

RB#25486