Bug #32163265: UBSAN ERROR WITH HYPERGRAPH OPTIMIZER: CONSTRUCTOR CALL ON MISALIGNED ADDRESS [noclose]
Hypergraph::Node had an alignment specification to be 64-aligned.
However, before C++17, there's no guarantee that new[] (and by
extension, std::vector<Node>) actually makes that alignment,
creating UBSan errors.

However, what it _did_ do is pad the object to be 64 bytes in size,
so we can do that instead and get the same result without the
UBSan errors. I've switched CPUs since the last measurements, and can no
longer reproduce the 60% speedup (it's down to 5–10%), but I've verified
that we get the same speedup from padding as from the alignment
declaration, and that _actual_ alignment does not help significantly
beyond some added noise (Coffee Lake 3.2 GHz).

Resubmit: Fix an issue that would break compilation on MSVC with
debug STL.

Change-Id: I65faa602bccaa030a67e522b8a3d08d35639485e