BUG#32175157 CONFIG PARSER DOES NOT IDENTIFY WHEN TWO MGMD NODES ARE GIVEN SAME PORT
When 2 management nodes are started on same port no error message was
shown on the console output, but the error was written to logfile. This
reduced the usability of mgmd as it made users to check everytime in
the logs to find the reason why mgmd failed to start.

The current fix adds validation in the ConfigInfo.cpp that checks if
two management nodes are specified same host and validates that they
dont have the same portnumber. If the validation is failed error is
thrown and mgmd node is prevented from starting.
Config rule validate_unique_mgm_port() is added that validates for
unique port for mgmd nodes on same hosts.

Validate_unique_mgm_port() loops through all the mgmd nodes and fetches
the Hostname and portnumber of that node. Key is created with hostname
and portnumber to add into unordered_map and if the key is duplicate
it is prevented from addition and error is thrown.

Change-Id: Icba48c0de987ace44d6c8a39e90c2162445ac82b