Bug#32131022: Heap-use-after-free near Item_func_in::cleanup_arrays
The problem may occur when we have a CREATE TABLE statement inside
a procedure, and the table is part of an existing view.
When the table is created, we check whether the table is part of an
existing view, and we call open_views_and_update_metadata() to open
and check the view. As part of this, we call open_tables() to open and
expand the view definition. Now, open_tables() will usually shift to
use the embedding procedure or statement's mem_root, because the view
expansion is usually scoped to the entire statement.
But in this case, we only want to perform a quick expansion within
the current execution, and we cleanup and delete the expanded
view definition after the expansion. But when the view expansion is
allocated on statement memory allocation context, the structures also
become part of the statement memory, and when the statement is
finally deallocated, we will actually attempt a second deallocation.

The solution to this is to force open_tables() to operate within
the current execution memory allocation context. This is done
by temporarily substituting the current statement arena's mem_root
pointer with the execution context mem_root pointer, while
open_tables() is executing.

Reviewed by: Gleb Shchepa <gleb.shchepa@oracle.com>