Bug#31165802: TABLE MDL ACQUIRED USING INAPPROPRIATE SCHEMA NAME CHARACTER CASE
Sometimes, for InnoDB background threads, table MDL was acquired using
an inappropriate character case for the schema name part of the lock
key. This happened for lower_case_table_names == 2. In this situation,
both the table name and the schema name should be lowercased before
acquiring the lock. However, in some cases, only the table name was
lowercased. The result of this was that MDL did not protect the meta
data the way it should, and we had race conditions such as the one
reported here.

The fix by this patch is to lowercase the schema name part of the MDL
key. Addtitionally, this patch fixes a couple of cases here MDL was
released before the corresponding DD objects, and we also improve upon
the asserts checking the lock protection when acquiring DD objects.

Change-Id: I32e1fbdf8a88cc6838203dc319e3c2d026ebef6a