BUG#32150315: MYSQLTEST HIDES OUTPUT FROM FAILING 'PERL' AND 'EXEC' COMMANDS
Background:

mysqltest can execute perl scripts and invoke arbitrary processes,
using the 'perl' and 'exec' commands.  mysqltest invokes the
sub-process, reads what it writes to its stdouts, and caches all that
in memory.  Then, mysqltest waits for the sub-process to exit.  Later
it writes that output to the output log (or whatever is the correct
destination, depending on configuration).

Problem:

If the sub-process fails, mtr discarded the output. That's a pity,
since the output probably contains some insight into the reason for
the error. (Or if it doesn't, we sometimes would need to add debug
output to the Perl script.)

Fix:

Flush the cached text before exiting.

Notes:

The result file mysql-test/r/main.result is changed. All for the good:
all changes are because of the additional, and potentially useful debug
info. For example, in the chunk near after "BUG#11731", the output now
contains the result sets from some queries that were actually executed
before the failing statement. In a real server test this may contain
information to debug the issue.

RB:25578