Bug#32150145: 8.0.22+ RELEASE BUILD QUERY FAILS SILENTLY,               DEBUG ASSERTION `THD->IS_ERROR()'
While pushing condition down to a materialized derived table, when a
condition is partially pushed down, optimizer calls "fix_fields" for
the condition that will remain in the outer query block. But the successful
return from this fix_fields() call is interpreted as error resulting in
server exit later in debug builds.
In most cases, the condition that remains in the outer query block is
already fixed. So the call to fix_fields() would not happen. However for
some cases a query transformation would have added new conditions to the
WHERE condition. In such a case, we might make a call to fix_fields().

Solution:
Interpret the return from fix_fields() call correctly.

Reviewed By : Roy Lyseng <roy.lyseng@oracle.com>