Bug #32162329 BOOL JSON_VALUE: ASSERTION `!CURRENT_THD->IS_ERROR()' FAILED.
This is a follow-up patch to
    Bug #31856260 STILL SEEING A FEW ASSERTION `THIS_TYPE != ENUM_JSON_TYPE::J_ERROR' FAILED

The functions json_is_valid() and sql_scalar_to_json() both do things like:
      longlong i = arg->val_int();
      if (current_thd->is_error()) return true;

i.e. they assume that if an error has been raised, the input was
malformed and could not be converted to json.

The underlying problems are caused by functions like
Item_func_plus::real_op() which continue to evaluate all arguments
even if an error has been raised. The fix is to check
current_thd->is_error() and do an early return.

Change-Id: Id68e54e242ea1a4c321d6a0219fef9137e3881de