Bug#32193123: DIFFERENT RESULTSET WITH HYPERGRAPH_OPTIMIZER SWITCH ENABLED
Wrong results were seen from a query which performed aggregation and
where a non-aggregated expression referenced a column in a derived
table. (Correct results were seen if the query used the hypergraph
optimizer.)

This is a regression from bug#28711717, and relates to how the
expressions are sorted in sort_copy_funcs(). Before bug#28711771,
expressions that referenced other expressions in the select list via
an alias, were placed last to ensure that the expressions they
depended on were evaluated first. This caused problems for expressions
generated by the IN-to-EXISTS transformation, which could contain
references to expressions in the select list, but without aliases.
These expressions also needed to be placed last to make sure their
dependencies were evaluated first.

Bug#28711717 fixed the problem for the IN-to-EXISTS transformation by
changing the reordering of the expressions, so that every expression
that referenced another expression in the select list was put last.
This made sure the synthesized expressions from the IN-to-EXISTS
transformation were handled correctly. But it also reordered
non-synthesized expressions with non-alias references. The latter was
unnecessary, but believed to be harmless. Now it turns out that the
extra reordering of the non-synthesized expressions caused problems
after all, as manifested by the wrong query results seen in this bug.

The fix is to modify the reordering logic so that non-aliased
references are reordered only if they were synthesized by the
IN-to-EXISTS transformation.

Change-Id: I066141553233a7e57d461cf295b8f2cda4e1c2c6