Bug#32199035: INTEGRATION WITH HYPERGRAPH OPTIMIZER: CRASH IN FINDBESTQUERYPLAN
If a secondary engine rejects all plans that connect all the tables of
the query, an assertion is hit in FindBestQueryPlan().

FindBestQueryPlan() has code to detect the case where the secondary
engine has rejected enough sub-plans to make it impossible to
construct a full plan. But it doesn't handle the case where enough
sub-plans are accepted to consider at least one full plan, but where
all the full plans are rejected.

Improve the check to detect the case where a full plan has been
considered, but no full plan is accepted, and raise an error if it
happens.

Change-Id: Ibf5bf88c19ffbc4fcda6fddf514c576f6ff9e9a4