BUG#32148921 ENABLING REDO HANG AFTER IMPORT WITH REDO DISABLE
Problem :
---------
While enabling redo log, the request for checkpoint hangs as sync flush
request up to specific LSN keeps looping with zero page as target. When
the number of pages in a iteration is less than the number of buffer
pools, the expression n_pages/srv_buf_pool_instances evaluates to zero.

This is a regression of previous bug-fix in 8.0.23. bug#31994031:SYNC
FLUSH SHOULD PLAN MORE PAGES TO FLUSH AHEAD TO AVOID STALLS

In earlier versions, we had a check to use "srv_io_capacity" as minimum
target for sync flush, which is removed by the above bug-fix. The above
fix evaluates exact number of pages to be flushed to reach a target LSN
and perhaps the minimum bound was considered unnecessary. However, the
exact value of n_pages gets reset afterwards in an attempt to divide
the load equally among buffer pools.

Solution:
---------
Keep the minimum bound for sync flush as safe guard. Since the
calculation is not expected to be exact and there is no harm flushing a
little more in last stage of sync flush, it should be fine keeping it.

Reviewed-by: Pawel Olchawa <pawel.olchawa@oracle.com>

RB: 25534