Bug #32206756: ASSERTION `!CURRENT_THD->IS_ERROR()' FAILED IN SQL/ITEM_CMPFUNC.CC
Remove all the current_thd->is_error() asserts. They are real signs of
the contract being violated (calling a val_* function with an error
condition set), but there are just too many of them, and this test case
happens in index condition pushdown and thus needs InnoDB fixes.
To really handle these, we'd need an overhaul of our error reporting,
so for the time being, we're fine with erroneous computations being
stopped a little too late in the evaluation.

Change-Id: I8925d8db1fba980b3c1c22e30c9a9277496fac77