BUG#32175785 ADD NEW MDL WEIGHT FOR NODES WAITING ON COMMIT ORDER
Analysis
--------
When a circularity involving multi-threaded applier workers is found, we wish
that the workers are the threads that back off:

1) Consider a deadlock scenario where the cycle contains a commit order
   wait where worker 2 is waiting for worker 1 to commit. Before this
   bugfix, it is possible that worker 1 is chosen as a victim. That does
   not unblock any thread, since worker 2 still needs worker 1 to
   commit. In fact, the only way to unblock commit order wait is to
   commit. Therefore, the result is a livelock-like scenario, where worker
   1 retries the transaction until the maximum number of retries has been
   reached, and then gives up and stops the applier.  By forcing the worker
   that waits on commit order to rollback, we prevent this scenario from
   happening, thus removing cases where the applier threads stop
   unnecessarily.

2) Being easy and convenient to do it (because of the already available
   retry infra-structure builtin the replication applier). For instance, if
   the observed circularity includes a client connection and a worker, is
   far more convenient that we don't make the client connection thread to
   back off and, subsequentially, the client issued statement fail.

Fix
---
Create an MDL graph weight dedicated to commit order queue waits that has
lower weight than any other and set commit order queue wait tickets to have
such weight.

Extra-fix
---------
Also, while submitting the patch to PB2, realized that there was a memory
leak in `memory::Aligned_atomic` move semantics, a missed delete of a
previously allocated buffer. Since both fixes relate to the same worklog,
WL#13574, the fix for that is also included in this patch.

Reviewed-by: Sven Sandberg <sven.sandberg@oracle.com>
Reviewed-by: Dimitry Lenev <dmitry.lenev@oracle.com>
RB: 25518